chore(iast): move inner funcionts
