+++ ddtrace/appsec/_iast/_handlers.py
@@ -5,14 +5,21 @@ from wrapt import when_imported
 from wrapt import wrap_function_wrapper as _w
 
 from ddtrace.appsec._iast import _is_iast_enabled
+from ddtrace.appsec._iast._iast_request_context import in_iast_context
 from ddtrace.appsec._iast._metrics import _set_metric_iast_instrumented_source
 from ddtrace.appsec._iast._patch import _iast_instrument_starlette_request
 from ddtrace.appsec._iast._patch import _iast_instrument_starlette_request_body
 from ddtrace.appsec._iast._patch import _iast_instrument_starlette_url
 from ddtrace.appsec._iast._patch import _patched_dictionary
 from ddtrace.appsec._iast._patch import try_wrap_function_wrapper
+from ddtrace.appsec._iast._taint_tracking import OriginType
+from ddtrace.appsec._iast._taint_tracking import is_pyobject_tainted
+from ddtrace.appsec._iast._taint_tracking import origin_to_str
+from ddtrace.appsec._iast._taint_tracking import taint_pyobject
 from ddtrace.appsec._iast._taint_utils import taint_structure
 from ddtrace.internal.logger import get_logger
+
+from ._iast_request_context import is_iast_request_enabled
 
 
 MessageMapContainer = None
@@ -48,15 +55,9 @@ def _on_set_http_meta_iast(
 
 
 def _on_request_init(wrapped, instance, args, kwargs):
-    from ddtrace.appsec._iast._iast_request_context import in_iast_context
-
     wrapped(*args, **kwargs)
     if _is_iast_enabled() and in_iast_context():
         try:
-            from ddtrace.appsec._iast._taint_tracking import OriginType
-            from ddtrace.appsec._iast._taint_tracking import origin_to_str
-            from ddtrace.appsec._iast._taint_tracking import taint_pyobject
-
             instance.query_string = taint_pyobject(
                 pyobject=instance.query_string,
                 source_name=origin_to_str(OriginType.QUERY),
@@ -75,8 +76,6 @@ def _on_request_init(wrapped, instance, args, kwargs):
 
 def _on_flask_patch(flask_version):
     if _is_iast_enabled():
-        from ddtrace.appsec._iast._taint_tracking import OriginType
-
         try_wrap_function_wrapper(
             "werkzeug.datastructures",
             "Headers.items",
@@ -132,11 +131,7 @@ def _on_flask_patch(flask_version):
 
 
 def _on_wsgi_environ(wrapped, _instance, args, kwargs):
-    from ddtrace.appsec._iast._iast_request_context import in_iast_context
-
     if _is_iast_enabled() and args and in_iast_context():
-        from ddtrace.appsec._iast._taint_tracking import OriginType
-
         return wrapped(*((taint_structure(args[0], OriginType.HEADER_NAME, OriginType.HEADER),) + args[1:]), **kwargs)
 
     return wrapped(*args, **kwargs)
@@ -145,8 +140,6 @@ def _on_wsgi_environ(wrapped, _instance, args, kwargs):
 def _on_django_patch():
     if _is_iast_enabled():
         try:
-            from ddtrace.appsec._iast._taint_tracking import OriginType
-
             # we instrument those sources on _on_django_func_wrapped
             _set_metric_iast_instrumented_source(OriginType.HEADER_NAME)
             _set_metric_iast_instrumented_source(OriginType.HEADER)
@@ -169,15 +162,9 @@ def _on_django_patch():
 
 
 def _on_django_func_wrapped(fn_args, fn_kwargs, first_arg_expected_type, *_):
-    # If IAST is enabled and we're wrapping a Django view call, taint the kwargs (view's
+    # If IAST is enabled, and we're wrapping a Django view call, taint the kwargs (view's
     # path parameters)
     if _is_iast_enabled() and fn_args and isinstance(fn_args[0], first_arg_expected_type):
-        from ddtrace.appsec._iast._iast_request_context import in_iast_context
-        from ddtrace.appsec._iast._taint_tracking import OriginType  # noqa: F401
-        from ddtrace.appsec._iast._taint_tracking import is_pyobject_tainted
-        from ddtrace.appsec._iast._taint_tracking import origin_to_str
-        from ddtrace.appsec._iast._taint_tracking import taint_pyobject
-
         if not in_iast_context():
             return
 
@@ -243,9 +230,6 @@ def _on_django_func_wrapped(fn_args, fn_kwargs, first_arg_expected_type, *_):
 
 
 def _custom_protobuf_getattribute(self, name):
-    from ddtrace.appsec._iast._taint_tracking import OriginType
-    from ddtrace.appsec._iast._taint_tracking import taint_pyobject
-
     ret = type(self).__saved_getattr(self, name)
     if isinstance(ret, (str, bytes, bytearray)):
         ret = taint_pyobject(
@@ -295,9 +279,6 @@ def _on_grpc_response(message):
 
 def if_iast_taint_yield_tuple_for(origins, wrapped, instance, args, kwargs):
     if _is_iast_enabled():
-        from ._iast_request_context import is_iast_request_enabled
-        from ._taint_tracking import taint_pyobject
-
         if not is_iast_request_enabled():
             for key, value in wrapped(*args, **kwargs):
                 yield key, value
@@ -316,17 +297,11 @@ def if_iast_taint_yield_tuple_for(origins, wrapped, instance, args, kwargs):
 
 def if_iast_taint_returned_object_for(origin, wrapped, instance, args, kwargs):
     value = wrapped(*args, **kwargs)
-    from ._iast_request_context import is_iast_request_enabled
 
     if _is_iast_enabled() and is_iast_request_enabled():
         try:
-            from ._taint_tracking import is_pyobject_tainted
-            from ._taint_tracking import taint_pyobject
-
             if not is_pyobject_tainted(value):
                 name = str(args[0]) if len(args) else "http.request.body"
-                from ddtrace.appsec._iast._taint_tracking import OriginType
-
                 if origin == OriginType.HEADER and name.lower() in ["cookie", "cookies"]:
                     origin = OriginType.COOKIE
                 return taint_pyobject(pyobject=value, source_name=name, source_value=value, source_origin=origin)
@@ -336,8 +311,6 @@ def if_iast_taint_returned_object_for(origin, wrapped, instance, args, kwargs):
 
 
 def _on_iast_fastapi_patch():
-    from ddtrace.appsec._iast._taint_tracking import OriginType
-
     # Cookies sources
     try_wrap_function_wrapper(
         "starlette.requests",
